requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
coverage==7.2.2
Django==4.1.7
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.2
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
